gun/pistols: move mesh manipulation to libtrx
