+++ src/libtrx/game/gun/pistols.c
@@ -1,0 +1,27 @@
+#include "game/gun/pistols.h"
+
+#include "game/gun/common.h"
+#include "game/lara/common.h"
+#include "game/sound.h"
+
+void Gun_Pistols_DrawMeshes(const LARA_GUN_TYPE weapon_type)
+{
+    Gun_SetLaraHandLMesh(weapon_type);
+    Gun_SetLaraHandRMesh(weapon_type);
+    Gun_SetLaraHolsterLMesh(LGT_UNARMED);
+    Gun_SetLaraHolsterRMesh(LGT_UNARMED);
+}
+
+void Gun_Pistols_UndrawMeshLeft(const LARA_GUN_TYPE weapon_type)
+{
+    Gun_SetLaraHandLMesh(LGT_UNARMED);
+    Gun_SetLaraHolsterLMesh(weapon_type);
+    Sound_Effect(SFX_LARA_HOLSTER, &Lara_GetItem()->pos, SPM_NORMAL);
+}
+
+void Gun_Pistols_UndrawMeshRight(const LARA_GUN_TYPE weapon_type)
+{
+    Gun_SetLaraHandRMesh(LGT_UNARMED);
+    Gun_SetLaraHolsterRMesh(weapon_type);
+    Sound_Effect(SFX_LARA_HOLSTER, &Lara_GetItem()->pos, SPM_NORMAL);
+}
