+++ src/libtrx/include/libtrx/game/gun.h
@@ -2,5 +2,5 @@
 
 #include "./gun/common.h"
 #include "./gun/const.h"
-#include "./gun/vars.h"
 #include "./gun/control.h"
+#include "./gun/vars.h"
