+++ src/libtrx/include/libtrx/game/gun/pistols.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "../lara/types.h"
+
+void Gun_Pistols_DrawMeshes(LARA_GUN_TYPE weapon_type);
+void Gun_Pistols_UndrawMeshLeft(LARA_GUN_TYPE weapon_type);
+void Gun_Pistols_UndrawMeshRight(LARA_GUN_TYPE weapon_type);
