+++ src/tr2/game/gun/gun.c
@@ -12,6 +12,7 @@
 #include <libtrx/config.h>
 #include <libtrx/debug.h>
 #include <libtrx/game/camera.h>
+#include <libtrx/game/gun/pistols.h>
 #include <libtrx/game/lara.h>
 
 void Gun_Control(void)
